Implement custom inline math delimiter
Multimarkdowns handling of dollar signs makes cases
such as "the amount of money ($M$) required to ..."
impossible to write because math-mode dollars are
required to stand free, i.e. have a space before the
opening dollar and one after the closing dollar.  It
is now possible to write "the amount of money (•M•) ..."
instead.